clean: move ufac file
