+++ src/pyhees/section3_1_e.py
@@ -16,7 +16,7 @@ from jjjexperiment.logger import log_res
 from jjjexperiment.inputs.di_container import *
 from jjjexperiment.inputs.app_config import *
 
-from jjjexperiment.underfloor_ac.inputs.common import UnderfloorAc
+from jjjexperiment.underfloor_ac.inputs.common import UnderfloorAc, UfVarsDataFrame
 
 # ============================================================================
 # E.2 床下温度
